requirements/base.txt
Django==1.9
djangorestframework==3.3.1
pymongo==3.1.1
wheel==0.26.0
django-rest-swagger
